+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Improve the ray biasing code to avoid self-intersections during the SSR traversal
 - Update Pyramid Spot Light to better match emitted light volume.
 - Moved _XRViewConstants out of UnityPerPassStereo constant buffer to fix issues with PSSL
+- Removed GetPositionInput_Stereo() and single-pass (double-wide) rendering mode
 
 ## [6.6.0-preview] - 2019-04-01